이행요청 삭제 수정

+++ src/main/java/kr/co/bestiansoft/ebillservicekg/eas/workRequest/controller/WorkRequestController.java
@@ -39,7 +39,7 @@ public class WorkRequestController {
 
     @ApiOperation(value="delete work request", notes= "delete work request")
     @DeleteMapping("/eas/workRequest/{reqId}")
-    public ResponseEntity<CommonResponse> deleteWorkRequest(@PathVariable Long reqId) {
+    public ResponseEntity<CommonResponse> deleteWorkRequest(@PathVariable int reqId) {
         workRequestService.deleteWorkRequest(reqId);
         return new ResponseEntity<>(new CommonResponse(200, "OK" ), HttpStatus.OK);
     }
